requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorzero==2.0
contourpy==1.3.0
cycler==0.12.1
diffusers==0.31.0
distro==1.9.0
docopt==0.6.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gpiozero==2.0.1
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
importlib_metadata==8.5.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.13.3
networkx==3.4.2
num2words==0.5.13
numpy==2.1.3
openai==0.28.0
opencv-contrib-python-headless
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
serial==0.0.97
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
streamlit==1.40.0
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.10
urllib3==2.2.2
uvicorn==0.32.0
watchdog==5.0.3
websockets==12.0
Werkzeug==3.1.3
yarl==1.17.1
zipp==3.21.0
